backend/app/requirements.txt
requests
sqlalchemy
pymysql
python-multipart
pydantic
PyJWT  
Pillow
python-dotenv
uvicorn
fastapi